Refractor Code to Obey SOLID principles
